8293293: Move archive heap loading code out of heapShared.cpp
