+++ src/hotspot/share/classfile/compactHashtable.cpp
@@ -26,7 +26,6 @@
 #include "jvm.h"
 #include "cds/archiveBuilder.hpp"
 #include "cds/cds_globals.hpp"
-#include "cds/heapShared.inline.hpp"
 #include "classfile/compactHashtable.hpp"
 #include "classfile/javaClasses.hpp"
 #include "logging/logMessage.hpp"
